move regression tests to a .test.ts file
